Added melons, movement graphics
+ Melons spawn randomly throughout the field and look like slices of watermelon.
+ Movement graphics is a triangle that changes with heading and cooldown
~ Not working: collision detection

+++ src/evolutionsim/Egg.java
@@ -2,7 +2,7 @@ package evolutionsim;
 
 public abstract class Egg extends Puck {
 
-    Egg(int[] dnaValues, double xcor, double ycor, double mass) {
+    Egg(double[] dnaValues, double xcor, double ycor, double mass) {
         super(dnaValues, xcor, ycor, mass);
         //randomizeGenes(dnaValues);
     }
